split bike and bikeparts into 2 apps with individual Dockerfiles
remove unneeded files

Add more detail in docker compose file
