+++ bikeparts/Dockerfile
@@ -6,7 +6,7 @@ WORKDIR /usr/src/app
 # Install app dependencies
 # A wildcard is used to ensure both package.json AND package-lock.json are copied
 # where available (npm@5+)
-COPY ../package*.json .
+COPY package*.json ./
 
 # If you are building your code for production
 # RUN npm ci --only=production
