+++ bikes/logger.js
@@ -1,0 +1,6 @@
+function log(req,res,next){
+    console.log(req.url);
+    next();
+}
+
+module.exports = log;
