+++ bikes/package.json
@@ -1,0 +1,15 @@
+{
+  "name": "bikes-app",
+  "version": "1.0.0",
+  "description": "Bikes microservices API",
+  "main": "BikeRoutes.js",
+  "scripts": {
+    "test": "echo \"Error: no test specified\" && exit 1"
+  },
+  "author": "",
+  "license": "ISC",
+  "dependencies": {
+    "express": "^4.17.1",
+    "joi": "^14.3.1"
+  }
+}
