--- package.json
@@ -1,19 +1,0 @@
-{
-  "name": "rest-demo",
-  "version": "1.0.0",
-  "description": "",
-  "main": "index.js",
-  "scripts": {
-    "test": "echo \"Error: no test specified\" && exit 1"
-  },
-  "keywords": [],
-  "author": "",
-  "license": "ISC",
-  "dependencies": {
-    "event-stream": "^4.0.1",
-    "express": "^4.16.3",
-    "joi": "^13.6.0",
-    "nodemon": "^2.0.4",
-    "typescript": "^3.0.3"
-  }
-}
